style: extracted studentView service
