+++ common-lib/src/main/kotlin/com/github/heheteam/commonlib/api/StudentApi.kt
@@ -8,18 +8,17 @@ import com.github.heheteam.commonlib.ResolveError
 import com.github.heheteam.commonlib.Student
 import com.github.heheteam.commonlib.SubmissionInputRequest
 import com.github.heheteam.commonlib.interfaces.AssignmentId
-import com.github.heheteam.commonlib.interfaces.AssignmentStorage
 import com.github.heheteam.commonlib.interfaces.CourseId
 import com.github.heheteam.commonlib.interfaces.CourseStorage
 import com.github.heheteam.commonlib.interfaces.CourseTokenStorage
 import com.github.heheteam.commonlib.interfaces.ProblemGrade
-import com.github.heheteam.commonlib.interfaces.ProblemStorage
 import com.github.heheteam.commonlib.interfaces.StudentId
 import com.github.heheteam.commonlib.interfaces.StudentStorage
 import com.github.heheteam.commonlib.interfaces.TokenError
 import com.github.heheteam.commonlib.logic.AcademicWorkflowService
 import com.github.heheteam.commonlib.logic.PersonalDeadlinesService
 import com.github.heheteam.commonlib.logic.ScheduledMessageService
+import com.github.heheteam.commonlib.logic.StudentViewService
 import com.github.heheteam.commonlib.logic.SubmissionSendingResult
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.Ok
@@ -31,14 +30,13 @@ import kotlinx.datetime.LocalDateTime
 @Suppress("LongParameterList", "TooManyFunctions")
 class StudentApi
 internal constructor(
-  private val courseStorage: CourseStorage,
-  private val problemStorage: ProblemStorage,
-  private val assignmentStorage: AssignmentStorage,
   private val academicWorkflowService: AcademicWorkflowService,
   private val personalDeadlinesService: PersonalDeadlinesService,
+  private val scheduledMessageDeliveryService: ScheduledMessageService,
+  private val studentViewService: StudentViewService,
+  private val courseStorage: CourseStorage,
   private val studentStorage: StudentStorage,
   private val courseTokenStorage: CourseTokenStorage,
-  private val scheduledMessageDeliveryService: ScheduledMessageService,
 ) {
   suspend fun checkAndSendMessages(timestamp: LocalDateTime): Result<Unit, EduPlatformError> =
     scheduledMessageDeliveryService.checkAndSendMessages(timestamp)
@@ -49,13 +47,13 @@ internal constructor(
   ): Result<List<Pair<Problem, ProblemGrade>>, EduPlatformError> =
     academicWorkflowService.getGradingsForAssignment(assignmentId, studentId)
 
-  fun getAllCourses(): Result<List<Course>, EduPlatformError> = courseStorage.getCourses()
+  fun getAllCourses(): Result<List<Course>, EduPlatformError> = studentViewService.getAllCourses()
 
   fun getStudentCourses(studentId: StudentId): Result<List<Course>, EduPlatformError> =
-    courseStorage.getStudentCourses(studentId)
+    studentViewService.getStudentCourses(studentId)
 
   fun getCourseAssignments(courseId: CourseId): Result<List<Assignment>, EduPlatformError> =
-    assignmentStorage.getAssignmentsForCourse(courseId)
+    studentViewService.getCourseAssignments(courseId)
 
   suspend fun inputSubmission(
     submissionInputRequest: SubmissionInputRequest
@@ -64,7 +62,7 @@ internal constructor(
   fun getProblemsFromAssignment(
     assignmentId: AssignmentId
   ): Result<List<Problem>, EduPlatformError> =
-    problemStorage.getProblemsFromAssignment(assignmentId)
+    studentViewService.getProblemsFromAssignment(assignmentId)
 
   fun loginByTgId(tgId: UserId): Result<Student, ResolveError<UserId>> =
     studentStorage.resolveByTgId(tgId)
